Add _raise_errors property to LDAPLoginManager (default False)
Add functionality to allow for LDAPLoginManager to raise errors on
not successfully finding a user instead of returning None.
It will return a different error message for if no users were found
vs. mismatch of username password.

Also changed the check utility to print out the specific error that
was received.