Get MAP_KEY results when logged in as actual user
Some LDAP systems let the bind_user see all attributes of another user, but some LDAP systems restrict what can be seen by the bind_user, especially if doing an anonymous bind. To make the MAP_KEY work as expected, re-search after connecting with the real user/password.

As an optimization, if a customer is happy with the results from the initial search, give them a config parameter to avoid the second search. This gives them functionality identical to today in case they need it. But it's opt-in to the old style. Doing nothing will give them the new behavior with MAP_KEY results when logged in as the actual user.